通常弾の移動追加
