+++ Assets/Battle/Bullet/Scripts/Base.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: bee9a4102fd370a4cba7d2f4c2e41cd7
+folderAsset: yes
+timeCreated: 1435917959
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
