+++ Assets/Battle/Bullet/Scripts/Base/BulletBasePalametar.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ce03649f94e2c074fb6ed844392c1dde
-timeCreated: 1435912538
+guid: b217c6d2a428cff438e4fb49cc8a56ab
+timeCreated: 1435917938
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
