+++ Assets/Battle/Bullet/Scripts/Bullet.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8c9889c489343b54094142542f8d5907
+folderAsset: yes
+timeCreated: 1435917809
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
