+++ Assets/Battle/Bullet/Scripts/Launcher.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9732148937c59bd4980788cd6f67ef77
+folderAsset: yes
+timeCreated: 1435918007
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
