+++ Assets/Battle/Bullet/Scripts/Missile/MissilePalametar.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 447f91d7f96b92846b31c5aefbd70d8d
+timeCreated: 1435918176
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
